# Speed up distributed Dask by removing reference to array_container in FITSLoader