--- a/dkist/io/loaders.py
+++ b/dkist/io/loaders.py
@@ -50,7 +50,7 @@ class BaseFITSLoader(metaclass=abc.ABCMeta):
         self.shape = shape
         self.dtype = dtype
         self.target = target
-        self.array_container = array_container
+        self.basepath = array_container.basepath
         self.ndim = len(self.shape)
         self.size = np.prod(self.shape)
 
@@ -67,10 +67,6 @@ class BaseFITSLoader(metaclass=abc.ABCMeta):
     @abc.abstractmethod
     def __getitem__(self, slc):
         pass
-
-    @property
-    def basepath(self):
-        return self.array_container.basepath
 
     @property
     def absolute_uri(self):