@Cadair 2024-10-28
Thanks a lot for your PR @svank :rocket: 

This is a great change, we haven't really had the time to do any significant optimisation on the loading code, so anything you can find would be very welcome.

I was curious to see the comparison between using distributed and using the [local processes](https://docs.dask.org/en/stable/scheduling.html#local-processes) scheduler, so I checked out your branch and loaded your data.

<details>
<summary>Setup</summary>

```python
import dask
from multiprocessing import Pool
import dkist
from astropy.io import fits
import numpy as np

ds = dkist.load_dataset("~/dkist_data/pid_2_86/BXXDZ")

ds_rebin = ds.rebin((1, 1, 936, 1))

def comp_slice(file):
    data = fits.getdata(ds.files.basepath / file)
    data = data[0]
    return data.sum(axis=0)
```
</details>

```python
# baseline serial, numpy
%timeit -n 1 -r 1 np.stack(list(map(comp_slice, ds.files.filenames)))
# 1min 57s ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)
```

```python
# simple multi-processing
p = Pool(16)
%timeit -n 1 -r 1 np.stack(p.map(comp_slice, ds.files.filenames))
# 12.3 s ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)
p.close()
```
```python
# Processes

dask.config.set(scheduler='processes')

%timeit -n 1 -r 1 ds.data.sum(axis=-2).compute();
# 18.1 s ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)

%timeit -n 1 -r 1 ds_rebin.data.compute()
# 18.3 s ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)
```

it's quite close, certainly good enough imo.

Without this PR, the last example took: `4min 34s ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)` so this looks to be a massive improvement to me.