@svank 2024-10-29
Interesting to see! With this PR I also have the `processes` scheduler coming in a bit slower than DIY parallelism, but twice as fast as the distributed scheduler. I wonder why---distributed's task stream has gaps but it's still mostly full. Maybe it has more overhead around starting and stopping each task?