@Cadair 2024-10-29
I suspect there's more overhead, but I don't know the details. 2x as fast without distributed is certainly eyebrow raising.