@Cadair 2024-10-29
@astrofrog suggested to me that it might be something to do with differences in how local process and distributed do or do not use threads on each worker.