@astrofrog 2024-10-29
As suggested to @Cadair right now I think the issue is that the chunks are very small and take <100ms to run, so distributed is going to be inefficient. I think it would be worth trying calling ``.rechunk`` to see if this helps.